pyproject.toml
[tool.poetry]
name = "arctic-calendar"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [
    { include = "arctic_calendar", from = "src" },
]

[tool.poetry.scripts]
arctic-calendar = "arctic_calendar.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^0.20.28"
patito = "^0.6.1"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
ruff = "^0.4.4"

[tool.ruff]
target-version = "py312"
line-length = 79

[tool.ruff.lint]
select = [
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # isort
    "I",
]

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
